tests/wpt/web-platform-tests/infrastructure/testdriver/set_permission.https.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>TestDriver set_permission method</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script>
    const descriptor = { name: "geolocation" };

    promise_test(async (t) => {
        await test_driver.set_permission(descriptor, "granted");
        permission = await navigator.permissions.query(descriptor);
        assert_equals(permission.state, "granted");
    }, "Grant Permission");

    promise_test(async (t) => {
        await test_driver.set_permission(descriptor, "denied");
        permission = await navigator.permissions.query(descriptor);
        assert_equals(permission.state, "denied");
    }, "Deny Permission");
</script>
